Yet more refactoring
